front page up and running
