+++ src/App.css
@@ -1,9 +1,8 @@
 .App {
-  text-align: center;
+  text-align: left;
 }
 
 .App-logo {
-  animation: App-logo-spin infinite 20s linear;
   height: 40vmin;
 }
 
@@ -14,19 +13,27 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
   color: white;
 }
 
-.App-link {
-  color: #61dafb;
+.main-title {
+  font-family: "October Crow";
+  letter-spacing: 1vmin;
+  font-size: 13vmin;
+  margin-top: -2vmin;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+#start-button {
+  font-size: 50px;
+  background-color: transparent;
+  border: none;
+  padding: 10px 30px;
+  color: white;
+  font-family: "October Crow";
+  line-height: 55px;
+}
+
+#start-button:hover {
+  font-size: 55px;
+  color: grey;
 }
