+++ src/fonts/october_crow/October Crow Font Info and Read Me.txt
@@ -1,0 +1,12 @@
+October Crow Font by Chad Savage
+http://www.sinisterfonts.com
+
+Created 2010, hand-lettered while watching The Crow on Devil's Night 2010. Turned into a PC font on Halloween, 2010.
+
+This is a free font, and I mean FREE. Use it however you want, for whatever you want. If you make something cool with it, I'd love to see it - email me at savage@sinistervisions.com and tell me about it. If you offer it for download on your own font site, PLEASE include this Read Me. That's not too much to ask, is it?
+
+This font is intended for specialty and graphic uses. It will look just awful if used in a paragraph format.
+
+For more free fonts and other downloadable goodies, visit http://www.sinisterfonts.com.
+
+FOR INSTRUCTIONS ON HOW TO INSTALL/USE THIS FONT, visit http://www.sinisterfonts.com, or just look it up on Google. Seriously. Don't email me and ask - this information is easily found with a 20-second search.
